Rename "gaugeColors" constant to "thresholds" and move to shared
