+++ ui/src/shared/components/Gauge.js
@@ -7,7 +7,7 @@ import {
   COLOR_TYPE_MIN,
   COLOR_TYPE_MAX,
   MIN_THRESHOLDS,
-} from 'src/dashboards/constants/gaugeColors'
+} from 'shared/constants/thresholds'
 
 class Gauge extends Component {
   constructor(props) {
